Fix grammar to handle indexStep only CFIs
These are CFIs that simply have a single path and a single step. For example: /4
I don't recall the details but my take is that the first step of a CFI expression is special and is known as the "indexStep" in the grammar.
A "path" expression in the grammar required both the "indexStep" and the continued paths in "localPath".
So, with the example CFI, there is only the indexStep and the localPath should be picked up as empty.
The grammar didn't support this so it would cause a SyntaxError with peg.js
See issue #72